requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
pymongo==4.6.3
motor==3.3.2
pydantic==1.10.22
python-dotenv==1.0.1
email-validator==1.3.1
aiohttp==3.8.5